refactor: extract camera logic to custom hook
