+++ src/shared/components/QuaiPayCamera/QuaiPayCamera.types.ts
@@ -1,0 +1,3 @@
+export enum ScannerType {
+  SEND_AMOUNT = 'SEND_AMOUNT',
+}
