+++ src/shared/components/index.tsx
@@ -1,5 +1,5 @@
 export { QuaiPayBanner } from './QuaiPayBanner';
-export { QuaiPayCamera } from './QuaiPayCamera';
+export { QuaiPayCamera } from './QuaiPayCamera/QuaiPayCamera';
 export { QuaiPayCard, CardSize } from './QuaiPayCard';
 export { QuaiPayContent } from './QuaiPayContent';
 export { QuaiPayInputDisplay } from './QuaiPayInputDisplay';
@@ -9,3 +9,6 @@ export { QuaiPayLoader } from './QuaiPayLoader';
 export { QuaiPayQRCode } from './QuaiPayQRCode';
 export { QuaiPaySearchbar } from './QuaiPaySearchbar';
 export { QuaiPayText } from './QuaiPayText';
+
+export { useQuaiPayCamera } from './QuaiPayCamera/QuaiPayCamera.hooks';
+export { ScannerType } from './QuaiPayCamera/QuaiPayCamera.types';
