+++ src/shared/hooks/index.ts
@@ -1,4 +1,5 @@
 export { useAmountInput } from './useAmountInput';
+export { useContacts } from './useContacts';
 export { useProfilePicture } from './useProfilePicture';
 export { useRetrieve } from './useRetrieve';
 export { useThemedStyle } from './useThemedStyle';
